Update README.md
Description: This is a Cab Booking Application. Using this application customers can book an inter-city cab. The driver 
object and vehicle object needs to have 4-5 records created initially. The cab reservation object requires all customer 
data and driver-cab preferences to be entered/selected. The automation used helps create the customer record and 
also sends emails to the customer and driver on the booking(record creation in the cab reservation object)..